splitting up base Model and Query
