+++ jetkit/db/soft_deletable.py
@@ -1,22 +1,5 @@
 from sqlalchemy import Column, DateTime, func
-from jetkit.db.query_filter import QueryFilter, FilteredQuery
-
-
-class SoftDeletableQueryFilter(QueryFilter):
-    """Omit rows marked as deleted."""
-
-    def apply_default_filter(self) -> "SoftDeletableQueryFilter":
-        assert isinstance(self, QueryFilter)
-        return self.filter(self.entity.deleted_at.is_(None))
-
-    def get_filter(self, obj: "SoftDeletable") -> bool:
-        return obj is None or not obj.deleted_at
-
-
-class SoftDeletableQuery(FilteredQuery):
-    """Query mixin."""
-
-    default_filters = [SoftDeletableQueryFilter]
+from jetkit.db.query.soft_deletable import SoftDeletableQuery
 
 
 class SoftDeletable:
@@ -26,3 +9,6 @@ class SoftDeletable:
 
     def mark_deleted(self) -> None:
         self.deleted_at = func.now()
+
+
+__all__ = ("SoftDeletableQuery",)
